make runner useable as a lib for other projects
Closes #59
