+++ package.json
@@ -68,6 +68,7 @@
       "real-user-agents",
       "tls-server",
       "runner",
+      "stacks",
       "../ulixee/hero/build/*",
       "../ulixee/hero/build/plugins/*"
     ]
