+++ runner/interfaces/userAgent.ts
@@ -1,0 +1,7 @@
+interface UserAgentConfig {
+    browserIds: string[];
+}
+
+export {
+    UserAgentConfig,
+};
