+++ runner/package.json
@@ -1,36 +1,15 @@
 {
   "name": "@double-agent/runner",
   "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "downloadRunnerData": "yarn workspace @double-agent/external-data downloadRunnerData",
-    "0": "node -r source-map-support/register scripts/0-collectFoundationalProfiles.js",
-    "1": "node -r source-map-support/register --max-old-space-size=10000 scripts/1-extractFoundationalProbes.js",
-    "2": "node -r source-map-support/register scripts/2-collectUserAgentsToTest.js",
-    "3": "node -r source-map-support/register scripts/3-runAssignments.js",
-    "3-hero": "yarn 3",
-    "3-secretAgent": "node -r source-map-support/register scripts/3-runAssignments.js --runner=secret-agent",
-    "3-puppeteer": "node -r source-map-support/register scripts/3-runAssignments.js --runner=puppeteer",
-    "4": "node -r source-map-support/register scripts/4-analyzeAssignmentResults.js"
-  },
   "dependencies": {
     "@double-agent/analyze": "1.0.0",
     "@double-agent/collect": "1.0.0",
     "@double-agent/collect-controller": "1.0.0",
     "@double-agent/real-user-agents": "1.0.0",
     "@double-agent/config": "1.0.0",
-    "@secret-agent/core": "^1.6.4",
-    "secret-agent": "^1.6.4",
-    "@ulixee/hero-fullstack": "^1.5.4",
-    "@ulixee/hero-core": "^1.5.4",
-    "@ulixee/chrome-97-0": "^4692.99.4",
     "node-fetch": "^2.6.1",
     "p-queue": "^6.6.2",
-    "puppeteer": "2.1.1",
     "unzipper": "^0.10.11",
     "commander": "^9.0.0"
-  },
-  "devDependencies": {
-    "@types/puppeteer": "2.0.0"
   }
 }
