+++ stacks/.gitignore
