+++ stacks/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@double-agent/stacks",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "downloadRunnerData": "yarn workspace @double-agent/external-data downloadRunnerData",
+    "0": "node -r source-map-support/register scripts/0-collectFoundationalProfiles.js",
+    "1": "node -r source-map-support/register --max-old-space-size=10000 scripts/1-extractFoundationalProbes.js",
+    "2": "node -r source-map-support/register scripts/2-collectUserAgentsToTest.js",
+    "3": "node -r source-map-support/register scripts/3-runAssignments.js",
+    "3-hero": "yarn 3",
+    "3-secretAgent": "node -r source-map-support/register scripts/3-runAssignments.js --runner=secret-agent",
+    "3-puppeteer": "node -r source-map-support/register scripts/3-runAssignments.js --runner=puppeteer",
+    "4": "node -r source-map-support/register scripts/4-analyzeAssignmentResults.js"
+  },
+  "dependencies": {
+    "@secret-agent/core": "^1.6.4",
+    "secret-agent": "^1.6.4",
+    "@ulixee/hero-fullstack": "^1.5.4",
+    "@ulixee/hero-core": "^1.5.4",
+    "@ulixee/chrome-97-0": "^4692.99.4",
+    "puppeteer": "2.1.1"
+  },
+  "devDependencies": {
+    "@types/puppeteer": "2.0.0"
+  }
+}
